Refactor SimpleQueue out of Crawler.
